🌍 #24 Tidy up Property code
- Moves GalaxyCheck.Property.Precondition => GalaxyCheck.Properties.Precondition
- Moves GalaxyCheck.TestResult => GalaxyCheck.Properties.TestResult
