+++ src/GalaxyCheck/Properties/ForAll.cs
@@ -8,14 +8,14 @@ namespace GalaxyCheck
         public static Property<object[]> Nullary(Func<bool> func) => new Property<object[]>(
             Gen.Constant(new object[] { }).Select(x => TestFactory.Create<object[]>(
                 x,
-                new Lazy<bool>(func),
+                func,
                 x)));
 
         public static Property<T0> ForAll<T0>(IGen<T0> gen0, Func<T0, bool> func) =>
             new Property<T0>(
                 gen0.Select(x => TestFactory.Create(
                     x,
-                    new Lazy<bool>(() => func(x)),
+                    () => func(x),
                     new object?[] { x })));
 
         public static Property<(T0, T1)> ForAll<T0, T1>(
@@ -24,7 +24,7 @@ namespace GalaxyCheck
             Func<T0, T1, bool> func) => new Property<(T0, T1)>(
                 Gen.Zip(gen0, gen1).Select(tuple => TestFactory.Create(
                     tuple,
-                    new Lazy<bool>(() => func(tuple.Item1, tuple.Item2)),
+                    () => func(tuple.Item1, tuple.Item2),
                     new object?[] { tuple.Item1, tuple.Item2 })));
 
         public static Property<(T0, T1, T2)> ForAll<T0, T1, T2>(
@@ -34,7 +34,7 @@ namespace GalaxyCheck
             Func<T0, T1, T2, bool> func) => new Property<(T0, T1, T2)>(
                 Gen.Zip(gen0, gen1, gen2).Select(tuple => TestFactory.Create(
                     tuple,
-                    new Lazy<bool>(() => func(tuple.Item1, tuple.Item2, tuple.Item3)),
+                    () => func(tuple.Item1, tuple.Item2, tuple.Item3),
                     new object?[] { tuple.Item1, tuple.Item2, tuple.Item3 })));
 
         public static Property<(T0, T1, T2, T3)> ForAll<T0, T1, T2, T3>(
@@ -45,7 +45,7 @@ namespace GalaxyCheck
             Func<T0, T1, T2, T3, bool> func) => new Property<(T0, T1, T2, T3)>(
                 Gen.Zip(gen0, gen1, gen2, gen3).Select(tuple => TestFactory.Create(
                     tuple,
-                    new Lazy<bool>(() => func(tuple.Item1, tuple.Item2, tuple.Item3, tuple.Item4)),
+                    () => func(tuple.Item1, tuple.Item2, tuple.Item3, tuple.Item4),
                     new object?[] { tuple.Item1, tuple.Item2, tuple.Item3, tuple.Item4 })));
 
         public static Property<object[]> Nullary(Action func) => Nullary(func.AsTrueFunc());
